Add SM process heartbeat and status to the hbs cluster
This update is the mtc hbsAgent side of a new
SM -> hbsAgent heartbeat algorithm for the
purpose of detecting peer SM process stalls.

This update adds an 'SM Heartbeat status' bit to
the cluster view it injects into its multicast
heartbeat requests.

Its peer is able to read this on-going hbsAgent/SM
heartbeat status through the cluster.

The status bit reads 'ok' while the hbsAgent sees
the SM heartbeat as steady.

The status bit reads 'not ok' while the SM heartbeat
is lost for longer than 800 msecs.

Change-Id: I0f2079b0fafd7bce0b97ee26d29899659d66f81d
Partial-Fix: #1895350
Co-Authored-By: Bin.Qian@windriver.com
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>