Remove 15 second delay following Swact Request status update
The Maintenance swact handler posts swact actions to the host
controller's task status field. The Swact Request posting was
followed by a 15 second wait so that Horizon would be
displaying "Swact: Request" while the swact occurred.

Unfortunately, this delayed the actual swact request for that
entire wait period thereby adding 15 seconds to the overall
manual swact operation.

Since it's better to run swact faster compared to waiting for
the status, this update removes that delay at the risk the
"Swact: Request" status not get displayed prior to the swact
taking place.

Change-Id: I635c896327dca2312efbe02dec67d3e920fa3e90
Closes-Bug: 1895767
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>