Fix AIO SX Lazy Reboot race condition
An race condition was found in the mtcClient's current
lazy reboot handling. The race condition is removed by
stopping pmond before entering lazy reboot wait loop.

A further enhancement was added to the mtcClient startup
process to detect if it had been restarted while acting
on a reboot request. In that unlikely case, the mtcClient
will now resume that action to ensure that the required
reboot occurs.

Change-Id: Ide1ba979c32e1d2005410e42ef6e0e14f10f5cb0
Closes-Bug: 1883519
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>