Fix reinstall of controller nodes
At shutdown, systemd will try to remount everything read-only
before attempting to unmount it. In the wipedisk script we
are deleting the partitions without unmounting
their corresponding filesystems. This leads to errors because
systemd will try to remount filesystems
whose partitions were deleted.

To fix this we have to unmount the filesystems that are linked to the
removed partitions.

Closes-Bug: 1919153
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>
Change-Id: I49a3c06ae6bce1324dd06f4fc63fb3e5cd4d28c1