Fix enabling heartbeat of self from the peer controller
This issue only occurs over an hbsAgent process restart
where the ready event response does not include the
heartbeat start of the peer controller.

This update reverts a small code change that was
introduced by the following update.

https://review.opendev.org/c/starlingx/metal/+/788495

Remove the my_hostname gate introduced at line 1267 of
mtcCtrlMsg.cpp because it prevents enabling heartbeat
of self by the peer controller.

Change-Id: Id72c35f25e2a5231a8a8363a35a81e042f00085e
Closes-Bug: 1922584
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>