Remove swerr log in hbsAgent cluster delete
The mtcAgent does not track the stopped or started
heartbeat state of a host, that is left to the
heartbeat service itself in response to the mtcAgent
commanding heartbeat start and stop based on current
running state.

Therefore heartbeat stop command is sometimes called
against a host that is already in the stopped state.

The heartbeat stop command results in a call in the
hbsAgent to delete a host from the heartbeat cluster;
hbs_cluster_del.

If that host is not already in the cluster then this
call can result in a Swerr (Software Error) log.

This update removes this success path Swerr log.

Change-Id: Idb96a791a932827749e329a123f60006ff7c48ec
Closes-Bug: 1931911
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>