Add crashkernel bootarg
kernel 5.10 did not support crashkernel=auto, so we change it to
crashkernel=512M, because kdump-anaconda-addon that is included in
squashfs.img does not support the value
--reserve-mb='2G-8G:256M,8G-16G:512M,16G-32G:1G,32G:2G'.

crashkernel=512M-2G:64M,2G-8G:256M,8G-16G:512M,16G-32G:1G,32G-:2G
that is equivalent to the 'auto' value for the 'crashkernel'.
This work will be done to make this dynamic, based on system memory,
via other tooling, such as puppet.

Story: 2008921
Task: 42872

Signed-off-by: Jiping Ma <jiping.ma2@windriver.com>
Change-Id: I9a12cf34074490fdfec897126bd9a32c530c36b1