Filter out kmod-bnxt_en{,-rt} as necessary
This commit ensures that the kmod-bnxt_en package is not installed when
in the low-latency profile, and that the kmod-bnxt_en-rt package is not
installed when in the standard/non-low-latency profile.

This is achieved by adding the kmod-bnxt_en and kmod-bnxt_en-rt packages
to the filter_out_from_* and pre_pkglist* files as required.

Testing:
- An ISO image can be built using a monolithic build procedure.
- Installation of the ISO image is successful with standard and
  low-latency profiles.
- The kmod-bnxt_en package is only installed with standard profile,
  whereas the kmod-bnxt_en-rt package is only installed with low-latency
  profile.

Story: 2009915
Task: 44792

Depends-On: https://review.opendev.org/c/starlingx/kernel/+/833464

Change-Id: I7a5c6086183d6b04e587192012f37aa7d0a2c2b3
Signed-off-by: M. Vefa Bicakci <vefa.bicakci@windriver.com>