Install packages from platform backup if prestage iso is absent
When the %pre code in post_miniboot_controller.cfg kickstart is run,
it checks if a prestage iso is present; if one is not, then it
attempts to download from the systemcontroller. The solution is
to add /opt/platform-backup/<release-id> as a repo of higher
priority than download, but lesser than prestage-iso,
in the %pre section in post_miniboot_controller.cfg.

Test Plan

PASS: Verify that packages are retrieved from
      /opt/platform-backup/<release-id> if they have been prestaged

PASS: Verify that packages are downloaded from systemcontroller
      for installation if packages are not prestaged in
      /opt/platform-backup/<release-id>

Closes-Bug: 1963709
Change-Id: Ibd2fe8d555c47da16e181b03703374bfd73731e3
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>