Precheck to avoid overwriting installed systems during prestaging
When booting the system with the prestaging ISO, check for the
presence of a valid installation.

Avoid overwriting an existing installation with a message to the
user.

Test Plan:

PASS: Verify that an existing installation is not overwritten by
      the prestaging operation

PASS: Verify failure handling of the prestaging operation with
      a message to the user, when an installation already exists
      on the subcloud.

PASS: Verify successful prestaging operation when there is no
      existing installation on the subcloud.

Story: 2009948
Task: 44872

Change-Id: I47e4f67f001da208f4e823ac699473c89b638ce5
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>