Debian KS: Populate patching ostree repo from ISO
To support the management of air gapped ostree patching, we need to also
install a copy of the ostree repo delivered with the ISO to a location
accessible to the patching subsystem.

This update integrates with the current PXE boot support for this
functionality.

Test Plan:
PASS - Build ISO
PASS - Boot/install ISO and confirm that the ostree_repo is copied
       successfully.
PASS - Confirm PXE booted system copies ostree_repo as expected

Change-Id: Id3a39d232c3d028caa6cbfe66fbfce7bbe2384c7
Story: 2009964
Task: 45015
Signed-off-by: Robert Church <robert.church@windriver.com>