Prestaging rootfs and bootfs customization with script provided via the prestage iso
When customizing the rootfs and bootfs, an external script is
provided via the prestage iso. The kickstart needs to execute
this script and obtain the rootfs and bootfs devices.

Test Plan:

PASS: Test the default prestaging operation on subclouds with
      different storage types (hard disk/ssd (nvme) )

PASS: Test the force install of prestage iso on different
      subclouds with different storage types with pre-existing
      installations

PASS: Test the graceful failure to install prestaging content
      when trying to install on subclouds (with different
      storage types) with pre-existing installations (and
      force-install is not used)

Regression:

PASS: Verify that autoinstaller installs successfully for all
      options

Story: 2009948
Task: 45048

Change-Id: Icc483d3430531a86c9a6488306a38e7d246ddd58
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>