Debian: Fix determination of http_port
http_port is derived from the install URL
however that URL is UNSET in the case of a USB or ISO
installation, so the default value of 8080 is used
in that scenario.

The http_port needs to be set in platform.conf before
bootstrap so that patching can be performed on a
system that has not been provisioned.

Test Plan:
  Build and initially boot a Debian ISO in vbox.
  Verify that the http_port is set to 8080.

Co-Authored-By: Eric MacDonald <eric.macdonald@windriver.com>
Story: 2009964
Task: 45161
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I847253d56f99bd95b72d5a8e960568cefa56c688