Debian KS: Update waiting for disks to become available
This work only affects Debian. This fixes an install issue.

With the introduction of [1] pre-start kickstart was split to
ks-early and pre-start. Logic was changed to use INSTDEV computed by
LAT between ks-early and pre-start. udevadm settle needs to happen
before computing INSTDEV, this means we need to move udevadm settle
from pre-start to ks-early.

[1] https://github.com/Wind-River/meta-lat/commit/
9cdcb08b9876e842d75ee7ba6e90bde1d67ac2c6

Tests on AIO-SX:
PASS: 10 times successful installs in a row, in an env that
      install failed ~6/10 times.

Story: 2009964
Task: 45068
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>
Change-Id: I5f37a3e2bc72c5f46171d8b3b794c21644b13482