Install package_checksums in /usr/local/share/pkg-list
In order to prepare shared software packages on the system
controller for subcloud prestaging, the file
"package-checksums" containing a list of packages and
their checksums, must be copied to /usr/local/share/pkg-list
and renamed as packages_list.

The prestaging process needs the package-and-checksum list
for the version of the load on the controller and the version
of the load on the subcloud.

This file needs to be copied over to /var/www/pages/feed/rel-<id>
on the controller so that the subclouds can download it.

This file can be obtained by downloading from the system
controller. If the system is being installed directly from
a boot iso, the appropriate kickstart is called to install
the file from the mounted bootimage.iso.

There are a few cases which determine how to get the file.

Cases:
1. Install load directly from bootimage.iso (both controller and
   subcloud) - copy from mounted iso to location. The iso
   may be present on a peripheral device like a CD/DVD.

2. Install load over lan - download the file from server and copy
   to locations (all controllers)

3. Version upgrade by importing a new load - copy the file
   to the appropriate location on the system controller.
   The subcloud will download the package_checksums belonging
   to the imported load as part of the upgrade process.

Test Plan:
PASS: Verify that the file is copied over to /usr/local/share/
      pkg-list and /var/www/feed/rel-<rel-id>/ when installing
      System Controller or subcloud from bootimage.iso

PASS: Verify that the file is copied to the locations when
      installing controller-1 via pxeboot

PASS: Verify that the file is copied to the locations when upgrading
      the System Controller (using load-import)

PASS: Verify that the file is copied when adding a subcloud with
      dcmanager subcloud add

Depends On: https://review.opendev.org/c/starlingx/distcloud/+/841828
Story: 2009799
Task: 45376

Change-Id: If14b197999f2ae03e15fdc08ec27511ca1d41767
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>