Remote subcloud install failure in post-install script
When the kickstart tries to download packages_list from the
system controller, wget fails because the parameter to its
"-O" argument needs the full file path (including the file
name) as opposed to the full path to the directory that the
file resides in.

The argument has been fixed to use the full path name of the
file.

Test Plan:
PASS: Verify that the remote subcloud successfully downloads the
      file and installation passes.

Closes-Bug: 1976384

Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: I65b6bd66e62e7353e33f2694ca5bf99d413df836