Delete Platform Backup partition first when installing Prestage iso
The size of a subcloud's platform backup partition can be increased
from a minimum of 30GB using a persistent_size entry in the file
install_values.yaml.

However, a failure to install prestage.iso is observed when a subcloud
has a platform backup partition size greater than 30GB. The failure to
install the prestage iso occurs, because "parted" attempts, by default
to create a 30GB platform backup partition in the prestaging kickstart.
"parted" fails when it encounters a partition size greater than 30GB.

This fix addresses this issue by always deleting the partition before
prestaging. This allows parted to create the partition with the default
value of 30GB, irrespective of what existed there previously.

Test Plan:
PASS: Verify that prestage ISO installation succeeds when the
      partition size of Platform Backup is greater than 30GB.

Regression:
PASS: Verify that prestage ISO installation succeeds when the
      force-install option is provided to gen_prestage_iso.sh

PASS: Verify that prestage ISO installation succeeds when the
      force-install option is not provided to gen_prestage_iso.sh
      and there is no pre-existing installation on the subcloud.

PASS: Verify that prestage ISO installation fails when the
      force-install option is not provided to gen_prestage_iso.sh
      and there is a pre-existing installation on the subcloud.

Closes-Bug: 1977975

Change-Id: I34fbf7bc6fb550deaee26e5c9a71402611834aa8
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>