Debian: OStree remote not set properly in kickstart for pxeboot installs
This update removes an if condition around setting the OSTree remote
following the pull in the kickstart. This is done so that the ostree
remote is set properly not only for USB but also now PXEBoot installs.

Test Plan:

PASS: Verify the ostree remote path is setup leaving the kickstart for
      both USB and PXEBoot installs.

Closes-Bug: 1979004
Change-Id: I9aab0552bfdb12ac579b351aa5081ab6b4a6ab21