debian: Move LAT partitioning/formatting into the kickstarts
Enable full control of the partition scheme in the kickstart files so
that partitioning can be more easily aligned with previous releases and
can be adjusted for specific personalities without needing additional
LAT / LAT installer changes.

This moves the equivalent logic from LAT into the kickstarts. This will
be cleaned up and optimized with a followup commit.

Test Plan:
 PASS - Install/Bootstrap/Provision AIO-SX
 PASS - Install/Bootstrap/Provision Std Controller + workers (2+2)

Change-Id: Ie4abf4479b9e8043c20712053b86afdd25ee626e
Story: 2009303
Task: 46191
Depends-On: https://review.opendev.org/c/starlingx/integ/+/856481
Signed-off-by: Robert Church <robert.church@windriver.com>