Align miniboot.cfg with recent kickstart.cfg refactor
This is a complete refactoring of miniboot.cfg to align with recent
changes to kickstart.cfg.

This is an extensive change, based off the refactored kickstart.cfg.
Relevant code from previous miniboot.cfg has been kept where applicable,
However, for consistency with the full install, most behaviour is taken
from kickstart.cfg unless it conflicts with the miniboot install..

Disk partitioning and filesystem sizing behaviour is now the same as
kickstart.cfg
  - NOTE: there are issues with expanding partitions based on
    persistent_size. These will be handled in a different commit.

Additionally, there are some minor tweaks/fixes in kickstart.cfg
- the only notable entry is adding a missing report_failure_with_msg
  call when checking backup partition device

Review Notes:

The best way to examine the code changes in this commit is by reviewing
the diff between the current kickstart.cfg and miniboot.cfg. Changes are
made with the goal of preserving a relatively sane diff between the two
files. Going forward it should be easier to track relevant changes
between them to pull in new features or bug fixes.

Test Plan
PASS:
- subcloud add/bootstrap on Dell hardware
    - UEFI boot
- subcloud add/bootstrap on sushy emulator (requires downgraded rvmc for
  now due to errors from sushy)
    - non-UEFI boot
    - graphical / serial
- Ensure platform_backup partition is detected and preserved
- lowlatency (install values 4/5)
- IPv4 / IPv6
    - Note: IPv6 vlan support is not available yet. This will be done
      in a separate commit
- Additional scenarios:
    - Repeat above add/bootstrap tests using custom-generated initrd-mini
      (initrd-mini included in ISO is still in development)
    - repeat tests after wipedisk / wipedisk --include-backup

Story: 2010118
Task: 46345

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I88567423c8bb0cef36863bb9f4eabdaa8980476e