Debian: UEFI RT pxeboot installs need efi=runtime option
Add the efi=runtime kernel command line option for all lowlatency UEFI
installs so that the efibootmgr has access to EFI non-volatile
variables. LAT will run the efibootmgr as one of that last steps of the
ISO/pxeboot install to change the device boot order of the host.

When the PREEMPT_RT kernel is enable this is required otherwise we will
rely on the existing boot order of the system, often resulting in a
pxeboot loop.

As reference here is the difference in the std and rt kernel requiring
this change:

    $ diff linux-yocto-{std,rt}/drivers/firmware/efi/efi.c
    69c69
    < static bool disable_runtime;
    ---
    > static bool disable_runtime = IS_ENABLED(CONFIG_PREEMPT_RT);
    98a99,101
    >
    >       if (parse_option_str(str, "runtime"))
    >               disable_runtime = false;

Test Plan:
PASS - Added parameter on a UEFI RT worker install in a H/W lab and
       observe that the previously failing post install reboot now
       boots successfully.

Signed-off-by: Robert Church <robert.church@windriver.com>
Closes-Bug: #1990895
Change-Id: Iff41e1125bb5a6e27f7de92862da1ef4899de794