Subcloud install failing due to non-existing directory (/etc/pmon.d)
Debian sub cloud is failing to bootstrap at the "Register kubelet
with pmond" task due to a non-existing directory, /etc/pmon.d.

This change ensures that the directory /etc/pmon.d exists, and that
the required files for different traits are appropriately linked.

TEST PLAN:

PASS: Verify that /etc/pmon.d is created on the subcloud when
      deploying the subcloud

PASS: Verify that the "Register kublet with pmond" ansible task
      completes

Closes-Bug: 1991968

Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: I6d71fad61bd0fa3a58cc1ab4f2d33fb7a7ec1bc7