Fix bashate failure in zuul
This review allows this repo to pass zuul.

When tox is run locally it pulls in an older
bashate 0.6.0 but the zuul jobs are pulling in
the higher version.

Bashate 2.1.1 was releated Oct 6, 2022

Changed the upper constraints to allow developers
to pull in dependencies that are more aligned with zuul.

Fixed the new bashate error.
Also cleaned up the yamllint syntax.

Closes-Bug: 1991971
Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: I9cda349a20c63f9d222a3c3fc3645c5ceb4c2751