Install from prestaged ostree_repo if exists in /opt/platform-backup
To make installation more efficient, we need to prestage the
ostree_repo in /opt/platform-backup.

When we add a subcloud, the ostree_repo will not be fetched from
the System Controller, but from the prestaged ostree_repo, if it
exists.

Test Plan:
PASS: Verify that the ostree_repo is not pulled from the system
      controller but from /opt/platform-backup for installation.

Story: 2010120
Task: 46691
Change-Id: I6a00b28abe96f5d254a77331fc231cd9edac7bea
Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>