Prestage container images at /opt/platform-backup
To save on network resources, it is required to prestage the
ostree_repo and the container images. When installation of a
subcloud is done, it is preferred that the prestaged ostree
repo and the prestaged container images are used, instead of
downloading them.

The prestaged container images are copied over to
/opt/platform-backup from the mounted media containing the
prestage iso.

Test Plan:
PASS: Verify that the prestaged container images are copied to
      /opt/platform-backup/<release version>.

PASS: Verify that the ostree_repo is not pulled from the system
      controller but from /opt/platform-backup for installation.

Story: 2010120
Task: 46709

Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: Ie2276d65e44f51a36dfcff922afaf5a9bdd0cf89