Basic time sync if required on subcloud install
This commit avoids install/bootstrap issues when the hwclock
is very far out of date with the system controller.

If the hwclock is more than approximiately 20m different than the
system controller then we set hwclock based on the system date

LAT initializes the system date based on the 'instdate' boot parameter.
The instdate boot parameter is the timestamp applied when the miniboot
bootimage.iso file is created on the system controller. It is close
enough to avoid any major out-of-sync system clock on the subcloud.

Secondary change: Optimize the interface assignment prior to ostree pull
This was mentioned during review
https://review.opendev.org/c/starlingx/metal/+/861017
Rather than employ sleeps, the /sys/class/net/${mgmt_dev}/operstate
file is used to determine when the interface has settled.
This is much more efficient than sleeps. We timeout after 60s.

Test Plan:
PASS:
- Install with subcloud in relative sync with system controller
  (within seconds). No hwclock change is applied to subcloud.
- Install when subcloud hwclock is more than 20m out of sync with
  system controller.
    - Verify that the hwclock is updated on the subcloud before
      ostree pull is initiated.
    - Verify that system date is proper on the first post-miniboot
      boot into the ostree installation.
- Test interface assignment wait/timeout functionality before
  ostree pulls. This is done on hardware subclouds (success path)
  and in sushy subclouds where failure mode testing was done
  by simulating stuck inteface operstate values.

Closes-Bug: 1995643
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: Ieddc774f962878f3c7f5886148310b87d4ffddfe