Revert "Enable Multinode Subcloud in Distributed Cloud"
This reverts commit d09313ff0b527efdcfd2c03bdfb950eb1432be10.

While the code here itself is functionally correct and tested,
a download in the code is dependent on a location on the
active System Controller that is overridden by a drbd2 mount
on /var/www/pages/iso.

This drbd2 mount masks the pxeboot related files which were
placed there during System Controller installation.

Reverting this change until a resolution to the drbd mount on
/var/www/pages/iso on the active System Controller is resolved.

Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: Ie91fde9a09f693d133fa484782a7df28ffd29faf