Fix failure to set instdev parameter when we use ks-setup.cfg
When we create the prestage iso with an external script,
ks-setup.cfg, we may not provide the rootfs_device or
boot_device parameter. This is a valid scenario where these
parameters are defined in ks-setup.cfg. An installation failure
is observed in this case.

The cause of the failure is that the prestage code is handled in
a pre-part hook. This commit moves it to a ks-early hook.

In this commit, a provision for the execution of a custom script
named ks-addon.cfg is also made. This script is a bash script
that must execute in the last post hook.

Test Plan:
PASS: Verify that the installation succeeds when the rootfs
      and boot device parameters are only specified via
      the ks-setup.cfg.

PASS: Verify that the external script, ks-addon.cfg, is executed
      after the install and configurations are done.

PASS: Verify that the logs from the execution of ks-addon.cfg
      are present in kickstart.log.

Closes-bug: 1997305

Signed-off-by: Shrikumar Sharma <shrikumar.sharma@windriver.com>
Change-Id: Ica1735aef3ab457cf0609ebee6aac45671e97987