Ensure magic strings that are visible for libblkid are erased
In the case when the root disk partition table is wiped but individual
partitions are not wiped correctly, this will leave previous physical
volume metadata intact on the disk.

When a new LVM partition is created and assigned as a newly created
physical volume the old LVM metadata on the disk partition will prevent
the cgts-vg volume group from being created.

This update will wipe all the magic strings present in the new physical
volume partition established by the kickstart by executing 'wipefs -a'
prior to creating the cgts-vg.

Test Plan:
PASS - Successfully install an ISO with this change on a system that did
       not cleanup the LVM metatadata from a previous install. Log in to
       the installed system and confirm that the cgts-vg is properly
       configured.

Change-Id: I63f4235a27cb40a4283f0f4c34f63564a4f18cdd
Partial-Bug: #1998204
Signed-off-by: Robert Church <robert.church@windriver.com>